changed byte to short. added merge syntax postgresql

+++ src/main/java/net/pms/dlna/DLNAMediaInfo.java
@@ -165,7 +165,7 @@ public class DLNAMediaInfo implements Cloneable {
 	private volatile ImageInfo imageInfo = null;
 	private String mimeType;
 	private final ReentrantReadWriteLock referenceFrameCountLock = new ReentrantReadWriteLock();
-	private byte referenceFrameCount = -1;
+	private short referenceFrameCount = -1;
 
 	private final ReentrantReadWriteLock avcLevelLock = new ReentrantReadWriteLock();
 	private String avcLevel = null;
@@ -2702,7 +2702,7 @@ public class DLNAMediaInfo implements Cloneable {
 	/**
 	 * @return reference frame count for video stream or {@code -1} if not parsed.
 	 */
-	public byte getReferenceFrameCount() {
+	public short getReferenceFrameCount() {
 		referenceFrameCountLock.readLock().lock();
 		try {
 			return referenceFrameCount;
@@ -2716,7 +2716,7 @@ public class DLNAMediaInfo implements Cloneable {
 	 *
 	 * @param referenceFrameCount reference frame count.
 	 */
-	public void setReferenceFrameCount(byte referenceFrameCount) {
+	public void setReferenceFrameCount(short referenceFrameCount) {
 		if (referenceFrameCount < -1) {
 			throw new IllegalArgumentException("referenceFrameCount must be >= -1.");
 		}
